Renamed the IS-05 parameters and added a distinct parameter for the websocket url
Renamed the IS-05 parameters and added a distinct parameter for the websocket url.
event_address is now two parameters broker_topic and websocket_url.
event_rest_url is now rest_api_url.

This change was made in order to keep the IS-05 transport parameters agnostic of their use and generic/reusable